requirements.txt
Flask==1.1.1
sentence-transformers==1.1.0
faiss-cpu==1.7.0
flask_sqlalchemy
flask_cors
SQLAlchemy-serializer